refactor: fold the never-shipped path_query_push intermediate into v1
The issue-#690 guard landed as path_query_push v1 gated to GROVE_V4,
but V4 has not been released, so no version table ever selected it —
once the per-instance engine took the V4 slot as v2, v1 was dead code
kept alive only by a doctored-version test. Since the slot values are
internal until V4 ships, collapse instead of leaving a hole: the
per-instance engine (which carries the #690 guard) IS v1 now, GROVE_V4
maps path_query_push back to 1, and the dispatcher knows [0, 1].
