+++ grovedb/src/element/path_query_push/mod.rs
@@ -12,15 +12,17 @@
 //!   slot when `decrease_limit_on_range_with_no_sub_elements` is set — even
 //!   when the emptiness was caused by `offset` skipping rows that did match
 //!   (issue #690), so e.g. `limit=2, offset=1` can return a single element.
-//! * **[v1]** — historical intermediate (never shipped alone). An empty
-//!   inner result consumes a limit slot only when nothing was skipped
-//!   (`skipped == 0`), i.e. only true no-match emptiness is charged;
-//!   offset-consumed subqueries leave the outer limit untouched.
-//! * **[v2]** — `GROVE_V4`+. Carries v1's guard, serves per-instance
+//! * **[v1]** — `GROVE_V4`+. Charges an empty inner result only when
+//!   nothing was skipped (`skipped == 0`, issue #690 — offset-consumed
+//!   subqueries are pagination, not absence), serves per-instance
 //!   limits (`Query::limit`) and reconciles subquery descents by total
 //!   consumed budget (rows plus empty-subtree charges) instead of
 //!   returned rows, aligning the read path's global-limit accounting
 //!   with the prover's shared-counter accounting.
+//!
+//!   (An intermediate carrying only the #690 guard was once gated here
+//!   for `GROVE_V4`; since no grove version ever shipped it, it was
+//!   folded into this v1 rather than kept as a dead dispatch arm.)
 //!
 //! Proof generation rejects non-zero offsets and never calls
 //! `path_query_push`, so this gate has no proof surface.
@@ -31,12 +33,9 @@
 //!
 //! [v0]: self::v0
 //! [v1]: self::v1
-//! [v2]: self::v2
 
 mod v0;
 mod v1;
-mod v2;
 
 pub(crate) use v0::path_query_push_v0;
 pub(crate) use v1::path_query_push_v1;
-pub(crate) use v2::path_query_push_v2;
